Adjust clock and battery polling based on results
Summary:
When the VEET first connects, it sometimes won't return a valid battery result the first time it is polled, so we want to poll it more frequently until we get a valid result.

Similarly, there is no reason to poll the clock frequently once we've determined we're in sync, so drop the poll rate after that.

Simplest way to do this was to convert back to recursive setTimeout calls since we adjust based on the results of the function. I hate polling.

Reviewed By: conordickinson

Differential Revision: D71831256

fbshipit-source-id: ac1d0e144c8d605de252399bb508638880700888