Fix possible bug in clock polling
Summary: I can't reproduce the original problem, but sometimes the clock is reported as out of sync and it doesn't resync for a whole minute. This fixes one possible source: if there is no response from the GT command then it would switch to the slow poll, which is incorrect. Also added logging to confirm.

Differential Revision: D74184206

fbshipit-source-id: 101d32dd902110c4237f63e7deaf9ee3f52cdf64