Rewrote time sync logic
Summary:
Rather than storing the time on the veet, I calculate an offset when we get the time and store that. I don't take into account travel time for the messages (which is about 200ms round trip on my mac right now), but that's ok as long as we're within 5 seocnds.

I changed the display to use the PC's clock plus the offset so that it doesn't look like we're way off, and more importantly we don't display that the time is off if it's been more than 20 seocnds since an update.

Also added some better language to clarify.

Differential Revision: D75459390

fbshipit-source-id: 44ab31dc8b5aa2960d16dc607e95c07c493c605b