Open folder if highlighted file is missing
Summary:
As reported by Aaron, if you don't have a Sensor_Data.csv file, when the user clicks the locate button it tries to open that file but the electron shell implementation just does nothing if the path doesn't exist.

So, first I open the directory the file is in using openPath, and then ask it to select the file. If the file exists, it will use the same window and then select it. If the file does not exist, it does nothing but leaves the directory view open.

Reviewed By: archthegit

Differential Revision: D78576888

fbshipit-source-id: 704e99f4d3b93953aee4e3d7aa01ef34e2c241d8